+++ skopt/learning/gaussian_process/__init__.py
@@ -1,3 +1,5 @@
 from .gpr import GaussianProcessRegressor
+from .gpr_betaWarp import GaussianProcessRegressor_BetaWarped
 
-__all__ = ("GaussianProcessRegressor")
+__all__ = ("GaussianProcessRegressor",
+           "GaussianProcessRegressor_BetaWarped")
